Started setting up the collider management minutiae.
There's a surprising amount of complexity in just passing the info around.

Really happy I didn't create mutable references. having to duplicate this code would be a huge use of time.

Also realized that `Box<dyn>` is a huge pain to use. Wrapping things in a simple enum is way easier. So that's now how pretty much all of PhysicsSystem's public interface works.

Though, sadly, can't do that with the main arena allocators as I don't think there's an easy way to edit the contents of an enum without unwrapping it.