IT FINALLY WORKS!
I had an index that was off-by-one... So it was simulating objects colliding with exactly themselves.

The unittest caught it, and now all the movement works flawlessly.

This also includes a change to push the "adjust movements up to the collision time" before the collision restitution. Was aiming to make the restitution more accurate (in case the slight difference was causing the bug). But instead found this.

Oh, also turned the "ignore collisions where moving apart" code back on just in case. It's probably a red-herring (and the real bug was the index), but minus well leave it in.