Refactored Entity to store an Orientation.
This removes the need to drag one around the step() function. That's good as 90% of the time if you're modifying an Entity, you also need to be able to transform into/out of it's local orientation.

As such the most risky part of this change is the center of mass calculations. Now the center of mass is the "position" and the Entity's local space origin (i.e. what colliders are stored relative to), is stored as an offset from that "position". So the center of mass needs to be a bit careful to not move the local space origin when mass is added.

Also, now Orientation doesn't hide it's 'rotation' since Entity can't hide it. Still, have added warnings that should prefer to use the affect_with() and after_affected() functions when possible.