Fixed the logic to ignore collisions between entities moving apart.
It wasn't accounting for angular velocity adding to the mix. Now it does, and it uses the same code that the collision restitution uses.

Note the simulation still is a bit wrong. I can see that certain collisions can lead to a massive increase in angular momentum, which destabilizes the system. Some energy conservation isn't being observed. But it takes a while to see it happen locally.