Continuing to look into how energy is being added.
Tracked it down to just the collision response. So it's not an accumulated error. I've been staring at the equations for a while, and I can't see where they deviate from the official equations I see everywhere... So maybe a value being fed in is wrong? I'm betting the moment of inertia tensors.