Realized that the plane collider still broke.
In particular, if a hit was at t=0 or t=1, it's just be ignored. More generally this meant that it wasn't treating the sphere as a range of values to overlap against, it was treating it as just it's closest point. So it'd either: hit between zero and one or go past and then hit the next iteration.