Adding a quick fix for tetrahedra sliding into the ground.
The problem was that the collision point was at the corners, which didn't 100% counter-act gravity. So now this will try to get the average of all close collision points. Had to setup some special epsilon calculation as the usual `EPSILON` was too small.

Though this doesn't seem to have completely fixed the issue. Possibly need to setup a way to make entities "go to sleep" when they are at rest.