Setup a way to get entities to "go to sleep" if they have little energy.
The main purpose of this is to allow things to come to rest on top of eachother. This should prevent the test tetrahedra from slowly sliding into the ground because it completely freezes it out.

The energy is tested at the end of the steps, so it should account for all collisions handled.

NOTE: This makes infinite mass entities act like they're (always) asleep, though I'm going to see if that can be removed as it doesn't really make sense. It will prevent such entities from being allowed to move around through their velocities.