Hooked the point collision half of mesh-to-mesh collision.
It mostly works. But every so often, I see the tetrahedron fall through the ground mesh and I don't think it's because of the missing line-line checking.

Also, had to setup a hack to guess the normal direction based on having all of the other points "vote" depending on whether they're on the same side of the intersection point as the normal. Should probably replace it with something more sound. Perhaps only have adjacent points (i.e. the ones connected by edges to the collision point) vote?