Refined the sleep handling for infinite masses.
Previously, things could get stuck such that forcing something to move after it came to rest under gravity could lead to the entity never being able to collide with the ground that was preventing it from falling. This was because the neighbor sets got too expansive.

So now infinite masses have some special handling

* They're put to sleep if they're not moving.
* Since there's little chance they'll ever wake up, other things will be added to/removed from their neighbor sets if those entities go to sleep.
* Wake up calls won't propagate through infinite mass entities (it wouldn't make sense, they don't respond to collisions).

The above error now has a test that's passing.