Can't just put things to sleep if they have low energy for one instant.
Because the vertical velocity can zero out when at the top of a jump. So have to only sleep if things are at low energy for multiple step() calls.

Note: Could've alternatively fixed this by including "potential energy" in the sleep check (i.e. at the end of a step, use each force generator to estimate an object's potential energy by seeing how far it could fall). But decided that was probably overly complicated and a bit too computationally expensive. This solution is much smaller. So until it shows some core defficiency, I'm going to stick with it.